# app/api/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Draft XP API
  version: 1.0.0

paths:
  /users:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "201":
          description: User created
        "400":
          description: Bad request
        "500":
          description: Internal server error

  /users/{id}:
    get:
      summary: Get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User object
        "404":
          description: User not found
        "500":
          description: Internal server error

  /contests:
    get:
      summary: List available contests
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contests

  /contests/{id}:
    get:
      summary: Get contest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contest details

  /contests/{contestId}/user/{userId}:
    post:
      summary: Enter a contest
      parameters:
        - name: contestId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Contestant created

  /contests/{id}/draftable-athletes:
    get:
      summary: Get draftable athletes for a contest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of athletes

  /contestants/{id}/roster:
    post:
      summary: Submit drafted roster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                athleteIds:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Roster submitted

  /contests/{id}/game:
    get:
      summary: Get full contest state (polling)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contest state including contestants, event feed, and boosts

  # /contestants/{id}/boosts:
  #   post:
  #     summary: Buy a boost
  #     parameters:
  #       - name: id
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               boostId:
  #                 type: string
  #     responses:
  #       "201":
  #         description: Boost purchased

  # /contestants/{id}/stats:
  #   patch:
  #     summary: Spend XP to upgrade stat power
  #     parameters:
  #       - name: id
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               stat:
  #                 type: string
  #                 enum: [points, rebounds, assists]
  #               xpSpent:
  #                 type: integer
  #     responses:
  #       "200":
  #         description: Stat power updated

  /contests:
    post:
      summary: Create a new contest with specified games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - gameIds
                - league
              properties:
                name:
                  type: string
                  description: Name of the contest
                gameIds:
                  type: array
                  items:
                    type: string
                  description: List of game IDs to include in the contest
                league:
                  type: string
                  enum: [nba, nfl, nhl, mlb]
                  description: League type for the contest
      responses:
        "201":
          description: Contest created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  gameIds:
                    type: array
                    items:
                      type: string
        "400":
          description: Bad request (invalid game IDs, league mismatch, etc.)
        "404":
          description: One or more games not found
        "500":
          description: Internal server error